Put grep_log function to utils
Function grep_log() could be useful for searching string patterns even
when Server instance is not possible. For example a case [1] when we
need to start Tarantool instance with our own options.

Now for using grep_log one needs create an empty Server instance and
then call grep_log() for it. It looks ugly:

```lua
local server = t.Server:new({})
local res = server:grep_log(msg, 1024, {filename = file})
```

Patch moves grep_log() to utils, this allows using grep_log() even
Server instance is not used.

1. https://github.com/tarantool/tarantool/pull/8836

Follows up https://github.com/tarantool/tarantool/issues/6128
