+++ luatest/utils.lua
@@ -1,4 +1,6 @@
+local errno = require('errno')
 local digest = require('digest')
+local fio = require('fio')
 local fun = require('fun')
 local yaml = require('yaml')
 
@@ -184,6 +186,86 @@ end
 
 function utils.is_tarantool_binary(path)
     return path:find('^.*/tarantool[^/]*$') ~= nil
+end
+
+--- Search a string pattern in the a log file.
+--
+-- @string pattern String pattern to search in the log file.
+-- @number[opt] bytes_num Number of bytes to read from the log file.
+-- @tab[opt] opts
+-- @bool[opt] opts.reset Reset the result when `Tarantool %d+.%d+.%d+-.*%d+-g.*`
+--   pattern is found, which means that the server was restarted.
+--   Defaults to `true`.
+-- @string[opt] opts.filename Path to the the log file.
+--   Defaults to `box.cfg.log`.
+-- @return string|nil
+function utils.grep_log(pattern, bytes_num, opts)
+    local options = opts or {}
+    local reset = options.reset or true
+    local filename = options.filename
+    local file = fio.open(filename, {'O_RDONLY', 'O_NONBLOCK'})
+
+    local function fail(msg)
+        local err = errno.strerror()
+        if file ~= nil then
+            file:close()
+        end
+        error(string.format('%s: %s: %s', msg, filename, err))
+    end
+
+    if file == nil then
+        fail('Failed to open log file')
+    end
+
+    io.flush() -- attempt to flush stdout == log fd
+
+    local filesize = file:seek(0, 'SEEK_END')
+    if filesize == nil then
+        fail('Failed to get log file size')
+    end
+
+    local bytes = bytes_num or 65536 -- don't read the whole log -- it can be huge
+    bytes = bytes > filesize and filesize or bytes
+    if file:seek(-bytes, 'SEEK_END') == nil then
+        fail('Failed to seek log file')
+    end
+
+    local found, buf
+    repeat -- read file in chunks
+        local s = file:read(2048)
+        if s == nil then
+            fail('Failed to read log file')
+        end
+        local pos = 1
+        repeat -- split read string in lines
+            local endpos = string.find(s, '\n', pos)
+            endpos = endpos and endpos - 1 -- strip terminating \n
+            local line = string.sub(s, pos, endpos)
+            if endpos == nil and s ~= '' then
+                -- Line doesn't end with \n or EOF, append it to buffer
+                -- to be checked on next iteration.
+                buf = buf or {}
+                table.insert(buf, line)
+            else
+                if buf ~= nil then
+                    -- Prepend line with buffered data.
+                    table.insert(buf, line)
+                    line = table.concat(buf)
+                    buf = nil
+                end
+                if string.match(line, '> Tarantool %d+.%d+.%d+-.*%d+-g.*$') and reset then
+                    found = nil -- server was restarted, reset the result
+                else
+                    found = string.match(line, pattern) or found
+                end
+            end
+            pos = endpos and endpos + 2 -- jump to char after \n
+        until pos == nil
+    until s == ''
+
+    file:close()
+
+    return found
 end
 
 return utils
